pyproject.toml
[project]
name = "osw-chatbot"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb==0.5.0",
    "duckduckgo-search>=6.3.3",
    "fastapi>=0.115.2",
    "langchain-community>=0.3.19",
    "langchain-core>=0.3.45",
    "langchain-graphrag>=0.0.9",
    "langchain-ollama>=0.2.3",
    "langchain-openai>=0.3.9",
    "langchain-text-splitters>=0.3.6",
    "langgraph>=0.3.14",
    "llm-sandbox[docker]>=0.2.3",
    "osw>=0.33.0",
    "panel>1.4.0",
    "playwright>=1.51.0",
    "pydantic[email]>=2.11.1",
    "requests>=2.32.3",
    "ruff>=0.11.2",
    "textract-py3>=2.1.0",
    "tiktoken>=0.9.0",
    "uvicorn>=0.31.1",
]
